Create connect4.py
First Pirple python course project.
Again here without import anything because not introduced.

The assignment of the project is:
You'll want to draw the board, and allow two players to take turns placing their pieces on the board (but as you learned above, they can only do so by choosing a column, not a row). The first player to get 4 across or diagonal should win!

Normally the pieces would be red and black, but you can use X and O instead.