Hangman in python
This is the second project in Pirple. Is the hangman game.
This is the requirement of the project:
For this assignment, we want to play hangman in 2-player mode. The game should start by prompting player 1 to pick a word. Then the screen should clear itself so that player 2 can't see the word.
After the screen is clear, the "gallows" and the empty letter spaces should be drawn, and player 2 should be allowed to guess letters until they either win, or lose. As they choose correct letters, the letters should appear on the screen in place of the blank space (clear and redraw the whole screen). As they choose wrong letters, the "man" himself should come end up being drawn, piece by piece. How many guesses they get before losing is up to you (depending on how complicated of a man you want to draw).

Plus I had insert the next homework that was error handling.
So the requirement was to handling errors on previous project and never trust on user input.

That is the result.