Briscola card game in Python
This is the last project in the pirple cours of python language.

Inside the python application you can call a guide with rules and all stuff to play, but if that is not clear to you, you can read other in here:
https://en.wikipedia.org/wiki/Briscola
Thanks to Wikipedia.

Here is the project requirement from pirple:

Everyone has their favorite card game. What's yours? For this assignment, choose a card game (other than Blackjack), and turn it into a Python program. It doesn't matter if it's a 1-player game, or a 2 player game, or more! That's totally up to you. A few requirements:

1. It's got to be a card game (no board games, etc)

2. When the game starts up, you should ask for the players' names. And after they enter their names, your game should refer to them by name only. ("It's John's turn" instead of "It's player 1's turn).

3. At any point during the game, someone should be able to type "--help" to be taken to a screen where they can read the rules of the game and instructions for how to play. After they're done reading, they should be able to type "--resume" to go back to the game and pick up where they left off.